+++ src/translation/translations/en.json
@@ -598,9 +598,6 @@
     "walletConnect": "WalletConnect"
   },
   "voteSummary": {
-    "for": "For",
-    "against": "Against",
-    "abstain": "Abstain",
     "addresses": "{{length}} addresses",
     "votes": "Votes"
   },
